+++ src/wallets/solana/accounts/index.ts
@@ -1,0 +1,5 @@
+/**
+ * @module Wallets/Solana/Accounts
+ */
+
+export { type SolanaAccountData, toSolanaAccount } from './solanaAccount'