+++ src/wallets/solana/actions/index.ts
@@ -1,0 +1,6 @@
+/**
+ * @module Wallets/Solana/Actions
+ */
+
+export { signMessage } from './signMessage'
+export { signTransaction } from './signTransaction'