+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.7.1
+
+### Patch Changes
+
+- [#44](https://github.com/openfort-xyz/openfort-node/pull/44) [`c59124b217a20296358e181425be37e80b9e176d`](https://github.com/openfort-xyz/openfort-node/commit/c59124b217a20296358e181425be37e80b9e176d) Thanks [@jamalavedra](https://github.com/jamalavedra)! - pregenerate only through account endpoints
+
+## 0.7.0
+
+### Minor Changes
+
+- [#32](https://github.com/openfort-xyz/openfort-node/pull/32) [`105c4668d86686698d260822347dd873925b446f`](https://github.com/openfort-xyz/openfort-node/commit/105c4668d86686698d260822347dd873925b446f) Thanks [@jamalavedra](https://github.com/jamalavedra)! - include backend wallets and update account namespace interactions'
+
 ## 0.6.78
 
 ### Patch Changes